A Provider was created, to handle the Shopping List state. The provider holds a list of Shopping Items, and exposes methods for adding and removing items from it.
Moved Item List state away from the Home widget to this Item Provider.

The method for adding a new items was moved to the Provider.

Created a Wrapper widget to hold the home widget and the new ChangeNotifierProvider which is responsible for passing down the Provider state to the rest of the widget tree.

The widget tests were modified to take create the Wrapper widget, instead of the home widget, so that widget test can still function.

Items are deleted from the Shopping list when the user taps the Delete Icon for the Tile which display that item. This has implemented like this:

The ItemList now passes an index to each of it's ItemTiles.
When the onPressed()-callback is called for the Delete IconButton we pass call the Providers removeItem()-method and pass the Tiles index as a parameter. The provider can then remove the Item which correspond to that index.

A key was added for each Delete-button. The key is connected to the new Widget Tests which test the deletion-functionality.